requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-mongodb-engine==0.6.0
django-mongoengine-filter==0.3.5
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
djangotoolbox==1.8.0
dnspython==2.1.0
google==3.0.0
google-auth==1.24.0
gunicorn==20.0.4
idna==2.10
mongoengine==0.22.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.2
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
